refactor: reorganize chat client structure and implement base classes
- Moved synchronous and asynchronous chat client implementations into a new `clients` directory.
- Introduced abstract base classes for synchronous and asynchronous chat clients to promote code reuse.
- Updated imports in the main chat module to reflect the new structure.
- Added exception handling for client requests and responses.
- Defined message and response types using Pydantic models for better data validation and handling.
