+++ packages/ragbits-chat/src/ragbits/chat/clients/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from .async_client import AsyncRagbitsChatClient
+from .sync_client import RagbitsChatClient
+
+__all__ = [
+    "AsyncRagbitsChatClient",
+    "RagbitsChatClient",
+]
